created snake game
Snake game is one of the most famous games available on all types of devices and works on every operating system. Snakes can move in every direction in this game, e.g., left, right, up, and down; after taking the food, the length of the snake increases. Food of the Snake will be generated at a given interval of time in C++ language and graphics function